+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Change `Gos\Bundle\WebSocketBundle\Periodic\DoctrinePeriodicPing` to address the deprecation of `Doctrine\DBAL\Driver\PingableConnection`
 - Deprecate support for `Doctrine\DBAL\Driver\PingableConnection` implementations in `Gos\Bundle\WebSocketBundle\Periodic\DoctrinePeriodicPing`, in 4.0 `Doctrine\DBAL\Connection` instances will be required
+- Add aliases to bundle events to allow registering listeners using the FQCN
 
 ## 3.2.0 (2020-06-01)